move DetailView into containers directory
